Create MagicSystem
