+++ CMakeLists.txt
@@ -107,6 +107,7 @@ add_executable(alicia-server
         src/server/chat/AllChatDirector.cpp
         src/server/chat/PrivateChatDirector.cpp
         src/server/messenger/MessengerDirector.cpp
+        src/server/race/MagicSystem.cpp
         src/server/race/RaceInstance.cpp
         src/server/race/RaceNetworkHandler.cpp
         src/server/ranch/RanchDirector.cpp
